requirements.txt
numpy==1.14.5
tensorflow==1.8.0
theano==1.0.2
cntk==2.5.1
keras==2.2.0
pandas==0.23.0
spacy==2.0.11
git+https://github.com/albermax/innvestigate@81cea8d71ad5b973855271a896e875e820ded823
scikit-learn==0.19.1
torch==0.4.0
pytorch-nlp==0.3.5
